Shared.makeSections: better section splitting inside a div.
We now propagate the div's classes and key/value attributes to
all the section divs created inside it.

See #5761.

Thus

```
::: foo
:::
```

puts the class `foo` on the section divs for both A and B.

Note that this is only one possible solution. Another approach
would be to force the whole div inside a title-less section.
Not sure which is best.

Currently 3 beamer tests fail because of this change.